interview_emotion_detection/requirements.txt
absl-py==2.3.1
astunparse==1.6.3
cachetools==6.2.2
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.2.0
cycler==0.12.1
decorator==4.4.2
facenet-pytorch==2.6.0
fer==22.5.1
ffmpeg==1.4
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.59.2
fsspec==2025.9.0
gast==0.6.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.15.1
idna==3.10
imageio==2.31.1
imageio-ffmpeg==0.4.7
Jinja2==3.1.6
keras==2.15.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.7.2
mdurl==0.1.2
ml_dtypes==0.2.0
moviepy==1.0.3
mpmath==1.3.0
namex==0.1.0
networkx==3.5
numpy==1.26.4
oauthlib==3.3.1
opencv-contrib-python==4.11.0.86
opencv-python==4.8.1.78
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
Pillow==10.2.0
proglog==0.1.12
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.2
pyparsing==3.0.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
six==1.17.0
sympy==1.14.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-macos==2.15.0
tensorflow-metal==1.1.0
termcolor==3.1.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.14.2
